pyproject.toml
[project]
name = "ollama-chatbot"
version = "0.1.0"
description = "An AI assistant chatbot built with Ollama and Streamlit"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "streamlit>=1.32.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "PyPDF2>=3.0.1",
    "python-docx>=1.1.0",
    "langchain>=0.1.9",
    "langchain-community>=0.0.27",
    "pydantic>=2.6.1",
    "typing-extensions>=4.9.0",
    "sentence-transformers>=2.5.1",
    "faiss-cpu>=1.7.4",
    "huggingface-hub>=0.20.3"
]
requires-python = ">=3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "setup.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError"
] 